Use cache stub if app.cache-mode=false
