+++ auth/src/main/resources/application-dev.properties
@@ -1,6 +1,7 @@
 app.test-mode=true
 app.jwt-param=jwt_param
 app.jwt-param-enabled=false
+app.cache-mode=true
 
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.url=jdbc:postgresql://localhost:5432/${DBNAME}
