+++ auth/src/main/resources/application-docker.properties
@@ -12,6 +12,7 @@ spring.datasource.hikari.maximumPoolSize=${HIKARI_MAX_POOL_SIZE}
 spring.datasource.hikari.auto-commit=false
 
 app.test-mode=${TEST_MODE_ENABLED}
+app.cache-mode=${CACHE_MODE_ENABLED:true}
 app.jwt-param=${JWT_PARAM}
 app.jwt-param-enabled=${JWT_PARAM_ENABLED:false}
